+++ libs/core/src/components/card-linked/card-linked.styles.ts
@@ -1,0 +1,53 @@
+import { css } from 'lit'
+
+const style = css`
+  :host {
+    display: contents;
+  }
+
+  a {
+    display: flex;
+    flex-direction: column;
+    box-sizing: border-box;
+    border-color: var(--_border-color);
+    background-color: var(--_background-color);
+    color: var(--_color);
+    border-width: var(--gds-sys-space-5xs);
+    border-style: solid;
+    text-decoration: none;
+    cursor: pointer;
+    outline: solid var(--gds-sys-space-4xs) transparent;
+    outline-offset: var(--gds-sys-space-4xs);
+    transition-property: all;
+    transition-duration: var(--gds-sys-motion-duration-fastest);
+    transition-timing-function: var(--gds-sys-motion-easing-ease-in);
+
+    &:focus {
+      outline-color: var(--gds-sys-color-content-neutral-01);
+      outline-offset: var(--gds-sys-space-3xs);
+
+      &:not(:focus-visible) {
+        outline-color: transparent;
+      }
+    }
+
+    @media (pointer: fine) {
+      &:hover {
+        background-color: color-mix(
+          in srgb,
+          var(--gds-sys-color-l2-neutral-02),
+          var(--gds-sys-color-state-neutral-03)
+        );
+      }
+    }
+
+    &:active {
+      background-color: color-mix(
+        in srgb,
+        var(--gds-sys-color-l2-neutral-02),
+        var(--gds-sys-color-state-neutral-04)
+      );
+    }
+  }
+`
+export default style