+++ libs/core/src/components/grouped-list/grouped-list.styles.ts
@@ -1,0 +1,20 @@
+import { css } from 'lit'
+
+const style = css`
+  .gds-list-heading {
+    padding: var(--gds-sys-space-xs) var(--gds-sys-space-m);
+    background: var(--gds-sys-color-l3-neutral-02);
+    border-radius: var(--gds-sys-radius-s);
+  }
+
+  ::slotted([gds-element='gds-list-item']) {
+    padding: var(--gds-sys-space-xs) var(--gds-sys-space-m);
+    border-bottom-color: var(--gds-sys-color-border-subtle-02);
+    border-bottom-width: var(--gds-sys-space-5xs);
+    border-bottom-style: solid;
+    display: flex;
+    justify-content: space-between;
+    gap: var(--gds-sys-space-3xs);
+  }
+`
+export default style