+++ libs/core/src/components/icon/icons/sort-ascending.ts
@@ -1,0 +1,5 @@
+import { IconSortAscending } from './sort-ascending.component'
+
+IconSortAscending.define()
+
+export { IconSortAscending }