+++ libs/react/src/lib/icon/icons/IconChevronRightSmall.tsx
@@ -1,0 +1,1 @@
+export { IconChevronRightSmall } from '@sebgroup/green-core/react';